requirements.txt
requests
colorama
python-dotenv
httpx